requirements.txt
flask
camelot-py[cv]
pandas
gunicorn
werkzeug
